requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/test.txt requirements/base.in requirements/test.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via pytest
configparser==4.0.2       # via importlib-metadata
contextlib2==0.6.0.post1  # via importlib-metadata, zipp
coverage==5.0.3           # via pytest-cov
ddt==1.2.2
djangorestframework==3.6.4
funcsigs==1.0.2           # via mock, pytest
importlib-metadata==1.5.0  # via pluggy, pytest
mock==3.0.5
more-itertools==5.0.0     # via pytest
packaging==20.1           # via pytest
pathlib2==2.3.5           # via importlib-metadata, pytest, pytest-django
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pyparsing==2.4.6          # via packaging
pytest-cov==2.8.1
pytest==4.6.9             # via pytest-cov, pytest-django
pytz==2019.3              # via django
scandir==1.10.0           # via pathlib2
six==1.14.0               # via mock, more-itertools, packaging, pathlib2, pytest
wcwidth==0.1.8            # via pytest
zipp==1.2.0               # via importlib-metadata
